% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATE_estimation.R
\name{ATE_estimation}
\alias{ATE_estimation}
\title{Average Treatment Effect Estimation}
\usage{
ATE_estimation(
  d,
  beta_draws,
  haz_draws,
  partition,
  covariates,
  trt,
  trt_values,
  n = 1000
)
}
\arguments{
\item{d}{the data}

\item{beta_draws}{the beta coefficients from the posterior draws, with dimension number of posterior draws * the number of variables}

\item{haz_draws}{the baseline hazard rates from the posterior draws, with dimension number of posterior draws * the number of intervals}

\item{partition}{the time values for each partition}

\item{covariates}{the names of the covariates used in the model}

\item{trt}{the treatment variable name}

\item{trt_values}{the possible values of the treatment, and the length should be two: reference and treatment}

\item{n}{the number of prediction for each posterior draw; the default is 1000}
}
\description{
Estimate the average treatment effect (ATE) based on the predicted survival times
}
\examples{
# example demo
}
