% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesgcomp.R
\name{bayesgcomp}
\alias{bayesgcomp}
\title{Average Treatment Effect Estimation}
\usage{
bayesgcomp(
  bayeshaz_object,
  ref,
  B = 1000,
  estimand = "prob",
  t = NULL,
  threshold = NULL,
  ...
)
}
\arguments{
\item{bayeshaz_object}{an object of the class \code{bayeshaz} created by the \code{bayeshaz()} function}

\item{ref}{the reference value of the treatment, so it should be one of the treatment values}

\item{B}{the number of prediction for each posterior draw; the default is 1000}

\item{estimand}{the statistics in interest; the default is the posterior survival
probability at each value of t, \code{prob}. The other options are the median survival time, \code{median}, and
the Restricted mean survival time, \code{rmean}.}

\item{t}{optional; a numeric vector of time points at which users want to compute marginal
survival probabilities}

\item{threshold}{optional; the threshold used for Restricted mean survival time. If not provided,
it will use the maximum time of the study as the default
survival time as a special case of Restricted mean survival time}

\item{...}{other parameters for estimand function}
}
\value{
This function returns an object of class \code{ATE} that stores the following information:
\itemize{
\item \code{surv_ref}, the posterior estimates for the reference
\item \code{surv_trt}, the posterior estimates for the treatment
\item \code{ref}, the value of the reference treatment
\item \code{trt_values}, the possible values of the treatment
\item \code{estimand}, the estimand for the ATE
\item \code{ATE}, the difference between the marginal survival probability of the treatment and the reference
\item \code{t}, the timepoints used for calculating the ATE, specifically for posterior survival probabilities
\item \code{threshold}, the threshold used for calculating the Restricted mean survival time. It would be
\code{null} if the estimand is not \code{rmean}.
}
}
\description{
Takes as input results from bayeshaz() and runs g-computation and returns posteriors draws of the average difference in survival probabilities between the two treatments.
By default, survival probabilities and differences are calculated at the midpoints of each of the partitions specified in bayeshaz().
}
\details{
Takes as input a \code{bayeshaz} object generated by bayeshaz() and runs g-computation and returns posteriors draws of the average difference in survival probabilities between the two treatments.
Averaging is done over a Bayesian bootstrap draw of the baseline confounder distribution.

It is important to specify the correct reference level for this function and also other ATE-related
ones. Users can specify or change reference by \code{relevel} function or the \code{ref} argument in these
ATE-related functions. The reference value needs to match the type of treatment variable, and it
needs to belong to one of the treatment values.

In addition, users can use argument \code{t} to specify the time points where they want to estimate
the posterior draws of the average difference of the survival probabilities. If not specified,
the function will use the default time points - the midpoints stored in the \code{bayeshaz} object.

These ATE-related functions only considers binary treatments so it is not applicable for more than two treatment values, or the multiple levels need to be coerced into a
binary variable before creating the \code{bayeshaz} object.
}
\examples{
\dontrun{
# example demo
## Continued from ?bayeshaz
gcomp_res = bayesgcomp(post_draws_ar1_adj, ## bayeshaz output 
                       ref = 0, ## treatment reference group
                       B = 1000, ## monte carlo iterations in g-comp
                       estimand = "prob"
                       )
}
}
\references{
Ji, Han, and Oganisian, Arman. 2023.
"causalBETA: An R Package for Bayesian Semiparametric Causal Inference with
Event-Time Outcomes. \emph{arXiv:2310.12358 [Stat]}, October.
\url{http://arxiv.org/abs/2310.12358}.
}
