% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayespipeline.R
\name{bayespipeline}
\alias{bayespipeline}
\title{causaBETA pipeline}
\usage{
bayespipeline(
  d,
  reg_formula,
  A,
  model = "AR1",
  sigma = 3,
  num_partitions = 100,
  warmup = 1000,
  post_iter = 1000,
  chains = 1,
  ref,
  t = NULL,
  V = 1000,
  func = function(x) {
     mean(x > t[i])
 },
  ...
)
}
\arguments{
\item{d}{data, a data frame in survival format. Categorical variables should be
transformed into dummy variables}

\item{reg_formula}{a formula object that specifies the formula for the poisson regression.
This also decides the formula will be used the function to check positivity overlap.}

\item{A}{a character variable that specifies the name of the treatment}

\item{model}{a character variable that tells the stan model used to implement the Bayesian piece-wise exponential model,
default is "AR1" and the other option is "independent"}

\item{sigma}{a numeric variable as the user-defined standard deviation for beta coefficients prior, the default is 3}

\item{num_partitions}{a numeric variable as the number of partitions of the study time, the default is 100}

\item{warmup}{a numeric variable as the number of warmup in MCMC, the default is 1000}

\item{post_iter}{a numeric variable as the number of iterations to draw from the posterior, the default is 1000}

\item{ref}{the reference value of the treatment, so it should be one of the treatment values}

\item{t}{optional; a numeric vector of time points at which users want to compute marginal
survival probabilities}

\item{V}{the number of prediction for each posterior draw; the default is 1000}

\item{func}{the function to calculate the statistics in interest; the default is the posterior survival
probability at each value of t, \code{function(x){mean(x > t[i])}}}

\item{...}{additional arguments required for func}
}
\value{
This function returns a list of one object of class \code{bayesahz} and \code{ATE} storing the modeling results
and the ATE result respectively. For details regarding these two objects, check \code{help(bayeshaz)} and \code{help(bayesgcomp)}.

It also plots the result of ATE by calling \code{plot()} on the ATE object.
}
\description{
Conduct the causalBETA pipeline in a row: bayesian piece-wise exponential modeling and the following average treatment effect estimation
}
\details{
The function is the wrapper function of the main function \code{bayeshaz()} and \code{bayesgcomp} to complete the computation
part of the analysis. For details, check the documentation for these two functions by \code{help(bayeshaz)} and \code{help(bayesgcomp)}.
}
\examples{
# example demo
}
