% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ATE.R
\name{create_ATE}
\alias{create_ATE}
\title{Construct an \code{ATE} object to store the results from g-computation}
\usage{
create_ATE(surv_ref, surv_trt, ref, trt_values, ATE, t)
}
\arguments{
\item{surv_ref}{a \code{mcmc} object storing the marginal survival probability for the reference}

\item{surv_trt}{a \code{mcmc} object storing the marginal survival probability for the treatment}

\item{ref}{the value of the reference treatment}

\item{trt_values}{a vector for the possible values of the treatment variable (binary)}

\item{ATE}{a numeric matrix for the difference between the marginal survival probability
of the treatment and the reference}
}
\value{
It returns an object of class 'ATE'
}
\description{
Construct an \code{ATE} object to store the results from g-computation
}
