% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bayeshaz.R
\name{create_bayeshaz}
\alias{create_bayeshaz}
\title{Construct the Object for Bayesian Piece-wise Exponential Model}
\usage{
create_bayeshaz(
  data,
  formula,
  treatment,
  covariates,
  time,
  outcome,
  model,
  sigma,
  chains,
  partition,
  midpoint,
  haz_draws,
  beta_draws
)
}
\arguments{
\item{data}{data, a data frame in survival format}

\item{formula}{a formula object that specifies the formula for the poisson regression.
This also decides the formula will be used the function to check positivity overlap.}

\item{treatment}{a character variable that specifies the name of the treatment}

\item{covariates}{a character variable storing the name of the covariate(s)}

\item{time}{a character variables that specifies the name of the time variable}

\item{outcome}{a character variables that specifies the name of the outcome variable}

\item{model}{a character variable recording which mode of model was used}

\item{sigma}{a numeric variable as the user-defined standard deviation for beta coefficients prior}

\item{chains}{the number of chains for sampling, the default is 1}

\item{partition}{a numeric vector for the paritition}

\item{midpoint}{a numeric vector for the midpoints of intervals}

\item{haz_draws}{a \code{mcmc.list} for the baseline hazard rates from posterior draws}

\item{beta_draws}{a \code{mcmc.list} for the beta coefficients estimated from posterior draws}
}
\value{
It returns an object of class \code{bayeshaz}
}
\description{
Construct a \code{bayeshaz} object to store the information of the bayesian piece-wise
exponential model and the posterior parameters. This constructor function is hidden
from the users
}
