% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ATE.R
\name{plot.ATE}
\alias{plot.ATE}
\title{Plot Average Treatment Effect or Marginal Survival Probability Estimated for an \code{ATE} Object}
\usage{
\method{plot}{ATE}(
  ATE_object,
  mode = "ATE",
  col_ATE = "black",
  col_0 = rgb(0.9, 0.1, 0.1, 0.9),
  col_1 = rgb(0.1, 0.1, 0.9, 0.9),
  col_CI_ATE = rgb(0.5, 0.5, 0.5, 0.5),
  col_CI_0 = rgb(0.9, 0.5, 0.5, 0.4),
  col_CI_1 = rgb(0.5, 0.5, 0.9, 0.4),
  xlim = NULL,
  ylim = NULL,
  type = "o",
  pch = 20,
  cex = 0.5,
  lwd = 1.5,
  main = NULL,
  xlab = NULL,
  ylab = NULL,
  ...
)
}
\arguments{
\item{ATE_object}{ATE_object an object of the class \code{ATE} created by the \code{ATE_estimation()} function}

\item{mode}{Mode of the plot, the default is \code{"ATE"} to plot the ATE in the \code{ATE} object. The other choices are
\code{0}, \code{1}, and \code{c(0, 1)}, which plot the control, treatment, and both respectively.}

\item{col_ATE, col_0, col_1}{the colors for the ATE, control and treatment points and lines. The default are
black, red, and blue respectively}

\item{col_CI_ATE, col_CI_0, col_CI_1}{the colors for the confidence interval of ATE, control and treatment.
The default are semitransparent grey, red, and blue}

\item{...}{other other graphical parameters for the plot function. Default ones will be used if not provided.}
}
\description{
This function displays the information stored in the \code{ATE} object: it can plot the marginal survival probability of
a treatment option or overlay both, and it can plot the ATE estimated corresponding to the reference value,
depending on the plotting choices user provided. It has some default plotting parameters if users don't specify otherwise.
}
\examples{
# example demo
}
