% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.bayehaz.R
\name{plot.bayeshaz}
\alias{plot.bayeshaz}
\title{Plot Average Baseline Hazard Rates}
\usage{
\method{plot}{bayeshaz}(
  bayeshaz_object,
  col_hazard = "black",
  col_CI = rgb(0.5, 0.5, 0.5, 0.5),
  ...
)
}
\arguments{
\item{bayeshaz_object}{an object of the class \code{bayeshaz} created by the \code{bayeshaz()} function}

\item{col_hazard}{the color parameter for the baseline hazard points, default is \code{black}}

\item{col_CI}{the color parameter for the confidence intervals of the baseline hazard, default is semitransparent-grey}

\item{...}{other graphical parameters for the plot function. Default ones will be used if not provided.}
}
\description{
This function plots the baseline hazard rate at the midpoint for each interval,
and it also marks the number at risk at the bottom corresponding to the ticks of the time axis.
}
\details{
Plot the average baseline hazard rate from all posterior draws on all intervals from the given \code{bayeshaz} object.
}
\examples{
# example demo
}
