% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ATE.R
\name{plot_ATE}
\alias{plot_ATE}
\title{Plot Average Treatment Effect Estimated}
\usage{
plot_ATE(ATE_estimated)
}
\arguments{
\item{ATE_estimated}{the list with the format: t, time points; ATE, ATE estimated from many posterior draws
#' @examples
# example demo}
}
\description{
Plot the average treatment effect (ATE) estimated from `ATE_estimation` function
}
