% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bayeshaz.R
\name{plot_bayeshaz}
\alias{plot_bayeshaz}
\title{Plot Average Baseline Hazard Rates}
\usage{
plot_bayeshaz(hazard, partitions, xv, obs_time)
}
\arguments{
\item{hazard}{a vector containing baseline hazard rates from all posterior draws for each interval}

\item{partitions}{a vector of time values for partitions}

\item{xv}{a vector of middle points for intervals. It is an returned output from bayeshaz()}

\item{obs_time}{a vector of the observed time of all subjects before events or censoring}
}
\description{
Plot the average baseline hazard rate from all posterior draws on all intervals generated from the bayeshaz() function
}
\examples{
# example demo
}
