% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_marginal_surv.R
\name{plot_marginal_surv}
\alias{plot_marginal_surv}
\title{Marginal Survival Curves Predicted}
\usage{
plot_marginal_surv(ATE_estimated, trt_name, trt_values, trt_option = c(1, 2))
}
\arguments{
\item{ATE_estimated}{the list with the format: t, surv_prob_1, surv_prob_2, and ATE.}

\item{trt_name}{the name of the treatment variable}

\item{trt_values}{the value of the treatment variable, and the order should be the same as the ATE_estimation() function}

\item{trt_option}{select treatment 1 or 2 or both. Default is c(1,2)}
}
\description{
Plot the marginal survival curves that are based on the results from ATE_estimation() function.
If both treatments are chosen, the difference between two curves would be the average treatment effect (ATE) estimated. It could be plotted with plot_ATE() function
}
\examples{
# example demo
}
