% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.bayeshaz.R
\name{predict.bayeshaz}
\alias{predict.bayeshaz}
\title{Piece-wise Constant Hazard Distribution Prediction}
\usage{
\method{predict}{bayeshaz}(bayeshaz_object, x, n = 1000, func)
}
\arguments{
\item{bayeshaz_object}{an object of the class \code{bayeshaz} created by the \code{bayeshaz()} function}

\item{x}{the individual used for prediction, with the variables specified in the model}

\item{n}{the number of prediction for each posterior draw; the default is 1000}

\item{func}{the function specified to get the quantities of interest; a common choice is \code{mean} to get the posterior distribution of the expectation of survival time}
}
\description{
Predict quantities of interest using the posterior draws from the piecewise exponential model
}
\examples{
# example demo
}
