% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.haz.R
\name{predict.haz}
\alias{predict.haz}
\title{Piecewise constant hazard distribution prediction}
\usage{
\method{predict}{haz}(x, beta_draws, haz_draws, partition, n = 1000, func)
}
\arguments{
\item{x}{the individual used for prediction, with the variables specified in the model}

\item{beta_draws}{the beta coefficients from the posterior draws, with dimension number of posterior draws * the number of variables}

\item{haz_draws}{the baseline hazard rates from the posterior draws, with dimension number of posterior draws * the number of intervals}

\item{partition}{the time values for each partition}

\item{n}{the number of prediction for each posterior draw; the default is 1000}

\item{func}{the function specified to get the quantities of interest; a common choice is `mean` to get the posterior distribution of the expectation of survival time}
}
\description{
Predict quantities of interest using the posterior draws from the piecewise exponential model
}
\examples{
# example demo
}
