% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_haz.R
\name{predict_haz}
\alias{predict_haz}
\title{Prediction helper function}
\usage{
predict_haz(x, beta_draws, haz_draws, partition, n = 1000, func)
}
\arguments{
\item{x}{the individual used for prediction, with the variables specified in the model}

\item{beta_draws}{the beta coefficients from the posterior draws, with dimension number of posterior draws * the number of variables}

\item{haz_draws}{the baseline hazard rates from the posterior draws, with dimension number of posterior draws * the number of intervals}

\item{partition}{the time values for each partition}

\item{n}{the number of prediction for each posterior draw; the default is 1000}

\item{func}{the function specified to get the quantities of interest; a common choice is \code{mean} to get the posterior distribution of the expectation of survival time}
}
\value{
The posterior predictive quantity in interest
}
\description{
Predict quantities of interest using the posterior draws from the piece-wise exponential model
}
