% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relevel.ATE.R
\name{relevel.ATE}
\alias{relevel.ATE}
\title{Modify the Reference Value}
\usage{
\method{relevel}{ATE}(ATE_object, ref)
}
\arguments{
\item{ATE_object}{an object of the class \code{ATE} created by the \code{ATE_estimation()} function}

\item{ref}{the new reference value, so it should be one of the treatment values}
}
\value{
An object of \code{ATE} with the new reference
}
\description{
The reference value of the \code{ATE} class object is changed to be the argument \code{ref}. If the new value provided is the
other treatment, then this function will exchange the marginal survival probability between reference and treatment.
}
\details{
When the reference provided is the other treatment than the current \code{ref} in the \code{ATE} object,
this function swaps the marginal survival probabilities, but it doesn't recalculate them for computational
efficiency. For calculation with a different set of weights, a new \code{ATE} object needs to be generated from
the \code{ATE_estimation()} function.
}
\examples{
# example demo
}
