% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayeshaz.R
\name{bayeshaz}
\alias{bayeshaz}
\title{Bayesian Piece-Wise Exponential Model}
\usage{
bayeshaz(
  d,
  reg_formula,
  A,
  model = "AR1",
  sigma = 3,
  num_partitions = 100,
  warmup = 1000,
  post_iter = 1000
)
}
\arguments{
\item{d}{the data frame in survival format, i.e. processed by Surv function}

\item{reg_formula}{the formula for the poisson regression}

\item{A}{the name of the treatment}

\item{model}{the stan model used, default is "AR1", other options are "independent" and "beta"}

\item{sigma}{the user-defined sd for odds ratio prior, the default is 3, the same as the default model}

\item{num_partitions}{the number of partitions of the study time, the default is 100}

\item{warmup}{the number of warmup in MCMC, the default is 1000}

\item{post_iter}{the number of iterations to draw from the posterior, the default is 1000}
}
\description{
Perform a bayesian piece-wise exponential model on the given survival data
Use an equivalent poisson regression in MCMC
}
\examples{
# example demo
}
