% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ATE.R
\name{create_ATE}
\alias{create_ATE}
\title{Construct an \code{ATE} object to store the results from g-computation}
\usage{
create_ATE(surv_ref, surv_trt, ref, trt_values, ATE, estimand, t, threshold)
}
\arguments{
\item{surv_ref}{a numeric matrix for the posterior estimates for the reference}

\item{surv_trt}{a numeric matrix for the posterior estimates for the treatment}

\item{ref}{the value of the reference treatment}

\item{trt_values}{a vector for the possible values of the treatment variable (binary)}

\item{ATE}{a numeric matrix for the difference between the marginal survival probability
of the treatment and the reference}

\item{estimand}{the estimand for this ATE}

\item{t}{timepoints used in this ATE object, specifically for posterior survival probability}

\item{threshold}{optional; the threshold used for restrictive mean survival time. If not provided,
it will use the maximum time of the study as the default
survival time as a special case of restrictive mean survival time}
}
\value{
It returns an object of class 'ATE'
}
\description{
Construct an \code{ATE} object to store the results from g-computation
}
