% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ATE.R
\name{plot.ATE}
\alias{plot.ATE}
\title{Plot Average Treatment Effect or Marginal Survival Probability Estimated for an \code{ATE} Object}
\usage{
\method{plot}{ATE}(
  ATE_object,
  mode = "ATE",
  col_ATE = "black",
  col_0 = rgb(0.9, 0.1, 0.1, 0.9),
  col_1 = rgb(0.1, 0.1, 0.9, 0.9),
  col_CI_ATE = rgb(0.5, 0.5, 0.5, 0.5),
  col_CI_0 = rgb(0.9, 0.5, 0.5, 0.4),
  col_CI_1 = rgb(0.5, 0.5, 0.9, 0.4),
  level_CI,
  xlim = NULL,
  ylim = NULL,
  type = "o",
  pch = 20,
  cex = 0.5,
  lwd = 1.5,
  main = NULL,
  xlab = NULL,
  ylab = NULL,
  ...
)
}
\arguments{
\item{ATE_object}{ATE_object an object of the class \code{ATE} created by the \code{bayesgcomp()} function}

\item{mode}{Mode of the plot, the default is \code{"ATE"} to plot the ATE in the \code{ATE} object. The other choices are
\code{0}, \code{1}, and \code{c(0,1)}, which plot the control, treatment, and both respectively.}

\item{col_ATE, col_0, col_1}{the colors for the ATE, control and treatment points and lines. The default are
black, red, and blue respectively}

\item{col_CI_ATE, col_CI_0, col_CI_1}{the colors for the confidence interval of ATE, control and treatment.
The default are semitransparent grey, red, and blue}

\item{level_CI}{Credible interval level, for a specified value an equal-tailed, level_CI\% credible interval will be plotted which has ((1-level_CI*100)/2)\% posterior probability below and above the interval. E.g. level_CI=.95 (the default) plots a 95\\% credible interval.}

\item{...}{other other graphical parameters for the plot function. Default ones will be used if not provided.}
}
\description{
This function displays the information stored in the \code{ATE} object: it can plot the marginal survival probability of
a treatment option or overlay both, and it can plot the ATE estimated corresponding to the reference value,
depending on the plotting choices user provided. It has some default plotting parameters if users don't specify otherwise.
}
\details{
When the mode is \code{ATE}, this function plots the ATE and its 95\% CI, and the default color is black and grey.
The default range of y-axis is from -1 to 1. The default title is \verb{ATE Estimated Over Time}.

When the mode is \code{0} or \code{1}, this function plots the marginal survival probability of the control or treatment solely with
the 95\% CI. The default colors are red and blue, and the default title is \verb{Marginal Survival Curve for ...}.

When the mode is \code{c(0,1)} (or \code{c(1,0)}), this function plots the control and treatment together with a default legend.
The default title is \verb{Marginal Survival Curves for Both Treatments}.
}
\examples{
# example demo
}
