% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_positivity.R
\name{plot_positivity}
\alias{plot_positivity}
\title{Positivity Assumption Checking}
\usage{
plot_positivity(d, formula, treatment)
}
\arguments{
\item{d}{the data frame in survival format, i.e. processed by Surv function}

\item{formula}{the formula for the logistic regression for propensity score (a formula object in this version)}

\item{treatment}{the treatment variable name}
}
\description{
Plot the histogram of the proprensity score of treatment assignment conditional on the observed treatment. 
This plot can be used to check the positivity assumption for casual inference.
If the assumption is met, the mirrored histogram should look similar.
If there is an obvious shift in the distribution of histograms, the assumption is likely violated.
}
\examples{
# example demo
}
