% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_positivity.R
\name{plot_positivity}
\alias{plot_positivity}
\title{Positivity Assumption Checking}
\usage{
plot_positivity(bayeshaz_object, data, formula = NULL, breaks = "Scott")
}
\arguments{
\item{bayeshaz_object}{an object of the class \code{bayeshaz} created by the \code{bayeshaz()} function}

\item{data}{the data set used to generate \code{bayeshaz} object}

\item{formula}{an optional formula variable used in fitting the propensity score model}

\item{breaks}{the parameter set for breaks, default is \code{"Scott"}}
}
\description{
Plot the histogram of the estimated propensity score by observed treatment.
This plot can be used to check the positivity assumption for casual inference.
If the assumption is met, there should be good overlap in the histograms. Histograms that do not overlap may indicate near violations of positivity.
}
\details{
If the formula is not given, this function will constructs the formula
for propensity score model by removing all treatment related terms from
the survival model formula, while the outcome should be the treatment variable.
If users provide a formula, it should not contain treatment variable nor the variables
that are not used in the survival model.
}
\examples{
# example demo
}
