% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_survival.R
\name{plot_survival}
\alias{plot_survival}
\title{Plot survival curve predicted}
\usage{
plot_survival(x, beta_draws, haz_draws, partition, n = 1000)
}
\arguments{
\item{x}{the individual used for prediction, with the variables specified in the model}

\item{beta_draws}{the beta coefficients from the posterior draws, with dimension number of posterior draws * the number of variables}

\item{haz_draws}{the baseline hazard rates from the posterior draws, with dimension number of posterior draws * the number of intervals}

\item{partition}{the time values for each partition}

\item{n}{the number of prediction for each posterior draw; the default is 1000}
}
\description{
Plot the survival curve based on the predictions from the posterior draw
}
\examples{
# example demo
}
